fix: expose tree-shakeable v6 helpers
